requirements.txt
altgraph==0.17
astroid==2.3.3
autopep8==1.5.1
colorama==0.4.3
cx-Freeze==6.1
future==0.18.2
helpdev==0.6.10
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pefile==2019.4.18
Pillow==7.0.0
psutil==5.7.0
pycodestyle==2.5.0
PyInstaller==3.6
pylint==2.4.4
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyQt5-stubs==5.14.1.0
PyQtChart==5.14.0
PySide2==5.14.1
pywin32-ctypes==0.2.0
QDarkStyle==2.8
QtPy==1.9.0
shiboken2==5.14.1
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
xlrd==1.2.0
zipp==3.1.0
